Initial commit from initial attempt at a mining explosive.  The idea is the explosion will preserve all dropped blocks, rather than letting any get deleted.
When I try to launch this source in Minecraft (1.6.4), it blows the heap
when I start the game.  It appears that
MiningTNTBlock.onNeighborBlockChange() is getting triggered during startup,
despite me not actually having a chance to even place blocks yet.

So this initial commit puts the source online for getting help in IRC.