Added a dummy texture for the mining TNT.  It's just "WAT" in monochrome.
Had to do a lot of fussing to get it to render properly.  It's not
rendering the top and bottom independently yet, although it's coded for
it; I just need to add more textures for doing that.

Discovered I needed a renderer to make this all work, looked at vanilla
and Mekanism for inspiration.  Once I got the common proxy to actually
call the helper (:p), it started to do something more than be a flashing,
white cube.

The renderer is in the common proxy, but it shouldn't belong there in the
long term.  I guess I need a client proxy for setting up the renderer.